Split middlewares
