+++ middlewares/response.js
@@ -1,0 +1,11 @@
+const debug = require("debug")("mock-server:middleware:response");
+
+module.exports = ({ server, services }) => async (req, res, next) => {
+  debug("response");
+  const {
+    requestHttp: { request, response }
+  } = req;
+  const { status } = request;
+
+  res.status(status).send(response);
+};
